Dodanie ograniczeń wartości dla klasy PPointXY.

+++ PP_Interpret_Gcode_Library/pp_prog_gcode_resolver_strategy.cpp
@@ -102,7 +102,7 @@ void PProgGcodeResolverStrategy::resetInterpretSettings(){
 }
 
 
-void PProgGcodeResolverStrategy::writeG00Line(PPpoint<double> &point){
+void PProgGcodeResolverStrategy::writeG00Line(PPpoint<int> &point){
 
 			if(destDevice->isOpen()){
 				PTextStream out(destDevice);
